refactor: update amount formatting in AvailableAssetsTable to display four decimal places; consolidate useFundAgent logic into FundAgent component
